+++ tasks.rb
@@ -3467,6 +3467,9 @@ def set_hpxml_roofs(hpxml_file, hpxml)
     roof_types = [[HPXML::RoofTypeClayTile, HPXML::ColorLight],
                   [HPXML::RoofTypeMetal, HPXML::ColorReflective],
                   [HPXML::RoofTypeWoodShingles, HPXML::ColorDark]]
+    int_finish_types = [[HPXML::InteriorFinishGypsumBoard, 0.5],
+                        [HPXML::InteriorFinishPlaster, 0.5],
+                        [HPXML::InteriorFinishWood, 0.5]]
     hpxml.roofs.clear
     roof_types.each_with_index do |roof_type, i|
       hpxml.roofs.add(id: "Roof#{i + 1}",
@@ -3477,6 +3480,8 @@ def set_hpxml_roofs(hpxml_file, hpxml)
                       emittance: 0.92,
                       pitch: 6,
                       radiant_barrier: false,
+                      interior_finish_type: int_finish_types[i % int_finish_types.size][0],
+                      interior_finish_thickness: int_finish_types[i % int_finish_types.size][1],
                       insulation_assembly_r_value: 2.3)
     end
   elsif ['base-atticroof-flat.xml'].include? hpxml_file
@@ -3489,6 +3494,7 @@ def set_hpxml_roofs(hpxml_file, hpxml)
                     emittance: 0.92,
                     pitch: 0,
                     radiant_barrier: false,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 25.8)
   elsif ['base-atticroof-conditioned.xml'].include? hpxml_file
     hpxml.roofs.clear
@@ -3500,6 +3506,7 @@ def set_hpxml_roofs(hpxml_file, hpxml)
                     emittance: 0.92,
                     pitch: 6,
                     radiant_barrier: false,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 25.8)
     hpxml.roofs.add(id: 'RoofUncond',
                     interior_adjacent_to: HPXML::LocationAtticUnvented,
@@ -3515,6 +3522,7 @@ def set_hpxml_roofs(hpxml_file, hpxml)
   elsif ['base-atticroof-cathedral.xml'].include? hpxml_file
     hpxml.roofs[0].interior_adjacent_to = HPXML::LocationLivingSpace
     hpxml.roofs[0].insulation_assembly_r_value = 25.8
+    hpxml.roofs[0].interior_finish_type = HPXML::InteriorFinishGypsumBoard
   elsif ['base-enclosure-garage.xml',
          'base-foundation-basement-garage.xml'].include? hpxml_file
     hpxml.roofs[0].area += 671
@@ -3547,6 +3555,8 @@ def set_hpxml_roofs(hpxml_file, hpxml)
       roof.roof_color = HPXML::ColorLight
       roof.emittance = nil
       roof.radiant_barrier = nil
+      roof.interior_finish_type = nil
+      roof.interior_finish_thickness = nil
     end
   elsif ['invalid_files/invalid-input-parameters.xml'].include? hpxml_file
     hpxml.roofs[0].radiant_barrier_grade = 4
@@ -3728,6 +3738,8 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     azimuth: 0,
                     solar_absorptance: 0.6,
                     emittance: 0.9,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
+                    interior_finish_thickness: 0.5,
                     insulation_assembly_r_value: 11.76)
     hpxml.walls.add(id: 'WallEast',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3738,6 +3750,8 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     azimuth: 90,
                     solar_absorptance: 0.6,
                     emittance: 0.9,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
+                    interior_finish_thickness: 0.5,
                     insulation_assembly_r_value: 11.76)
     hpxml.walls.add(id: 'WallSouth',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3748,6 +3762,8 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     azimuth: 180,
                     solar_absorptance: 0.6,
                     emittance: 0.9,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
+                    interior_finish_thickness: 0.5,
                     insulation_assembly_r_value: 11.76)
     hpxml.walls.add(id: 'WallWest',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3758,6 +3774,8 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     azimuth: 270,
                     solar_absorptance: 0.6,
                     emittance: 0.9,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
+                    interior_finish_thickness: 0.5,
                     insulation_assembly_r_value: 11.76)
     hpxml.walls.add(id: 'WallAtticGableEast',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3803,6 +3821,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 1200,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallAtticGable',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3823,6 +3842,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 686,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallOther',
                     exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
@@ -3831,6 +3851,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 294,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 4.0)
   elsif ['base-bldgtype-single-family-attached.xml'].include? hpxml_file
     hpxml.walls.clear
@@ -3842,6 +3863,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 686,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallOther',
                     exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
@@ -3850,6 +3872,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 294,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 4.0)
     hpxml.walls.add(id: 'WallAtticGable',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3875,6 +3898,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
   elsif ['base-atticroof-cathedral.xml'].include? hpxml_file
     hpxml.walls[1].interior_adjacent_to = HPXML::LocationLivingSpace
     hpxml.walls[1].insulation_assembly_r_value = 23.0
+    hpxml.walls[1].interior_finish_type = HPXML::InteriorFinishGypsumBoard
   elsif ['base-atticroof-conditioned.xml'].include? hpxml_file
     hpxml.walls.delete_at(1)
     hpxml.walls.add(id: 'WallAtticKneeWall',
@@ -3884,6 +3908,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 316,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallAtticGableCond',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3893,6 +3918,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 240,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 22.3)
     hpxml.walls.add(id: 'WallAtticGableUncond',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -3912,6 +3938,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 100,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallOtherMultifamilyBufferSpace',
                     exterior_adjacent_to: HPXML::LocationOtherMultifamilyBufferSpace,
@@ -3920,6 +3947,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 100,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallOtherNonFreezingSpace',
                     exterior_adjacent_to: HPXML::LocationOtherNonFreezingSpace,
@@ -3928,6 +3956,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 100,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23.0)
     hpxml.walls.add(id: 'WallOtherHousingUnit',
                     exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
@@ -3936,6 +3965,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 100,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 4.0)
   elsif ['base-enclosure-walltypes.xml'].include? hpxml_file
     walls_map = { HPXML::WallTypeCMU => 12,
@@ -3955,6 +3985,12 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     [HPXML::SidingTypeStucco, HPXML::ColorLight],
                     [HPXML::SidingTypeVinyl, HPXML::ColorDark],
                     [HPXML::SidingTypeNone, HPXML::ColorMedium]]
+    int_finish_types = [[HPXML::InteriorFinishGypsumBoard, 0.5],
+                        [HPXML::InteriorFinishGypsumBoard, 1.0],
+                        [HPXML::InteriorFinishGypsumCompositeBoard, 0.5],
+                        [HPXML::InteriorFinishPlaster, 0.5],
+                        [HPXML::InteriorFinishWood, 0.5],
+                        [HPXML::InteriorFinishNone, nil]]
     last_wall = hpxml.walls[-1]
     hpxml.walls.clear
     walls_map.each_with_index do |(wall_type, assembly_r), i|
@@ -3966,6 +4002,8 @@ def set_hpxml_walls(hpxml_file, hpxml)
                       color: siding_types[i % siding_types.size][1],
                       area: 1200 / walls_map.size,
                       emittance: 0.92,
+                      interior_finish_type: int_finish_types[i % int_finish_types.size][0],
+                      interior_finish_thickness: int_finish_types[i % int_finish_types.size][1],
                       insulation_assembly_r_value: assembly_r)
     end
     hpxml.walls << last_wall
@@ -3981,6 +4019,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 2080,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23)
     hpxml.walls.add(id: 'WallGarageInterior',
                     exterior_adjacent_to: HPXML::LocationGarage,
@@ -3989,6 +4028,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 320,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23)
     hpxml.walls.add(id: 'WallGarageExterior',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -4018,6 +4058,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 960,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23)
     hpxml.walls.add(id: 'WallGarageInterior',
                     exterior_adjacent_to: HPXML::LocationGarage,
@@ -4026,6 +4067,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 240,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23)
     hpxml.walls.add(id: 'WallGarageExterior',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -4086,6 +4128,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 294,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 4.0)
   elsif ['invalid_files/invalid-assembly-effective-rvalue.xml'].include? hpxml_file
     hpxml.walls[0].insulation_assembly_r_value = 0
@@ -4095,6 +4138,8 @@ def set_hpxml_walls(hpxml_file, hpxml)
       wall.solar_absorptance = nil
       wall.color = HPXML::ColorMedium
       wall.emittance = nil
+      wall.interior_finish_type = nil
+      wall.interior_finish_thickness = nil
     end
   elsif ['base-foundation-basement-garage.xml'].include? hpxml_file
     hpxml.walls.add(id: 'WallGarageBasement',
@@ -4104,6 +4149,7 @@ def set_hpxml_walls(hpxml_file, hpxml)
                     area: 320,
                     solar_absorptance: 0.7,
                     emittance: 0.92,
+                    interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                     insulation_assembly_r_value: 23)
     hpxml.walls.add(id: 'WallGarageExterior',
                     exterior_adjacent_to: HPXML::LocationOutside,
@@ -4125,13 +4171,14 @@ end
 def set_hpxml_foundation_walls(hpxml_file, hpxml)
   if ['ASHRAE_Standard_140/L322XC.xml'].include? hpxml_file
     hpxml.foundation_walls.add(id: 'FoundationWallNorth',
-                               exterior_adjacent_to: 'ground',
+                               exterior_adjacent_to: HPXML::LocationGround,
                                interior_adjacent_to: HPXML::LocationBasementConditioned,
                                height: 7.25,
                                area: 413.25,
                                azimuth: 0,
                                thickness: 6,
                                depth_below_grade: 6.583,
+                               interior_finish_type: HPXML::InteriorFinishNone,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4139,13 +4186,14 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                insulation_exterior_distance_to_top: 0,
                                insulation_exterior_distance_to_bottom: 0)
     hpxml.foundation_walls.add(id: 'FoundationWallEast',
-                               exterior_adjacent_to: 'ground',
+                               exterior_adjacent_to: HPXML::LocationGround,
                                interior_adjacent_to: HPXML::LocationBasementConditioned,
                                height: 7.25,
                                area: 195.75,
                                azimuth: 90,
                                thickness: 6,
                                depth_below_grade: 6.583,
+                               interior_finish_type: HPXML::InteriorFinishNone,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4153,13 +4201,14 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                insulation_exterior_distance_to_top: 0,
                                insulation_exterior_distance_to_bottom: 0)
     hpxml.foundation_walls.add(id: 'FoundationWallSouth',
-                               exterior_adjacent_to: 'ground',
+                               exterior_adjacent_to: HPXML::LocationGround,
                                interior_adjacent_to: HPXML::LocationBasementConditioned,
                                height: 7.25,
                                area: 413.25,
                                azimuth: 180,
                                thickness: 6,
                                depth_below_grade: 6.583,
+                               interior_finish_type: HPXML::InteriorFinishNone,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4167,13 +4216,14 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                insulation_exterior_distance_to_top: 0,
                                insulation_exterior_distance_to_bottom: 0)
     hpxml.foundation_walls.add(id: 'FoundationWallWest',
-                               exterior_adjacent_to: 'ground',
+                               exterior_adjacent_to: HPXML::LocationGround,
                                interior_adjacent_to: HPXML::LocationBasementConditioned,
                                height: 7.25,
                                area: 195.75,
                                azimuth: 270,
                                thickness: 6,
                                depth_below_grade: 6.583,
+                               interior_finish_type: HPXML::InteriorFinishNone,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4185,6 +4235,8 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
       hpxml.foundation_walls[i].insulation_interior_r_value = 10.2
       hpxml.foundation_walls[i].insulation_interior_distance_to_top = 0.0
       hpxml.foundation_walls[i].insulation_interior_distance_to_bottom = 7.25
+      hpxml.foundation_walls[i].interior_finish_type = HPXML::InteriorFinishGypsumBoard
+      hpxml.foundation_walls[i].interior_finish_thickness = 0.5
     end
   elsif ['base.xml'].include? hpxml_file
     hpxml.foundation_walls.add(id: 'FoundationWall',
@@ -4194,6 +4246,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 1200,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4211,6 +4264,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 686,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4224,6 +4278,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 294,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4240,6 +4295,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
     hpxml.foundation_walls[0].interior_adjacent_to = HPXML::LocationBasementUnconditioned
     hpxml.foundation_walls[0].insulation_exterior_distance_to_bottom = 0
     hpxml.foundation_walls[0].insulation_exterior_r_value = 0
+    hpxml.foundation_walls[0].interior_finish_type = HPXML::InteriorFinishNone
   elsif ['base-foundation-unconditioned-basement-wall-insulation.xml'].include? hpxml_file
     hpxml.foundation_walls[0].insulation_exterior_distance_to_bottom = 4
     hpxml.foundation_walls[0].insulation_exterior_r_value = 8.9
@@ -4264,6 +4320,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
     hpxml.foundation_walls[0].area /= 2.0
     hpxml.foundation_walls[0].depth_below_grade -= 4
     hpxml.foundation_walls[0].insulation_exterior_distance_to_bottom -= 4
+    hpxml.foundation_walls[0].interior_finish_type = HPXML::InteriorFinishNone
   elsif ['base-foundation-multiple.xml'].include? hpxml_file
     hpxml.foundation_walls[0].area = 600
     hpxml.foundation_walls.add(id: 'FoundationWallInterior',
@@ -4304,6 +4361,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 480,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4317,6 +4375,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 120,
                                thickness: 8,
                                depth_below_grade: 3,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4330,6 +4389,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 60,
                                thickness: 8,
                                depth_below_grade: 1,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4345,6 +4405,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 160,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4358,6 +4419,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 240,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4371,6 +4433,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 160,
                                thickness: 8,
                                depth_below_grade: 3,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4384,6 +4447,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 120,
                                thickness: 8,
                                depth_below_grade: 3,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4397,6 +4461,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 80,
                                thickness: 8,
                                depth_below_grade: 3,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4421,6 +4486,8 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
   elsif ['base-misc-defaults.xml'].include? hpxml_file
     hpxml.foundation_walls.each do |fwall|
       fwall.thickness = nil
+      fwall.interior_finish_type = nil
+      fwall.interior_finish_thickness = nil
     end
   elsif ['invalid_files/invalid-facility-type-surfaces.xml'].include? hpxml_file
     hpxml.foundation_walls.add(id: 'FoundationWallOther',
@@ -4430,6 +4497,7 @@ def set_hpxml_foundation_walls(hpxml_file, hpxml)
                                area: 294,
                                thickness: 8,
                                depth_below_grade: 7,
+                               interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                                insulation_interior_r_value: 0,
                                insulation_interior_distance_to_top: 0,
                                insulation_interior_distance_to_bottom: 0,
@@ -4448,12 +4516,14 @@ end
 def set_hpxml_frame_floors(hpxml_file, hpxml)
   if ['ASHRAE_Standard_140/L100AC.xml',
       'ASHRAE_Standard_140/L100AL.xml'].include? hpxml_file
-    hpxml.frame_floors.add(id: 'FloorUnderAttic',
+    hpxml.frame_floors.add(id: 'CeilingBelowAttic',
                            exterior_adjacent_to: HPXML::LocationAtticVented,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 1539,
+                           interior_finish_type: HPXML::InteriorFinishGypsumBoard,
+                           interior_finish_thickness: 0.5,
                            insulation_assembly_r_value: 18.45)
-    hpxml.frame_floors.add(id: 'FloorOverFoundation',
+    hpxml.frame_floors.add(id: 'FloorAboveFoundation',
                            exterior_adjacent_to: HPXML::LocationOutside,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 1539,
@@ -4470,23 +4540,25 @@ def set_hpxml_frame_floors(hpxml_file, hpxml)
          'ASHRAE_Standard_140/L324XC.xml'].include? hpxml_file
     hpxml.frame_floors.delete_at(1)
   elsif ['base.xml'].include? hpxml_file
-    hpxml.frame_floors.add(id: 'FloorBelowAttic',
+    hpxml.frame_floors.add(id: 'CeilingBelowAttic',
                            exterior_adjacent_to: HPXML::LocationAtticUnvented,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 1350,
+                           interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                            insulation_assembly_r_value: 39.3)
   elsif ['base-bldgtype-multifamily.xml'].include? hpxml_file
     hpxml.frame_floors.clear
-    hpxml.frame_floors.add(id: 'FloorOther',
+    hpxml.frame_floors.add(id: 'FloorAboveOther',
                            exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 900,
                            insulation_assembly_r_value: 2.1,
                            other_space_above_or_below: HPXML::FrameFloorOtherSpaceBelow)
-    hpxml.frame_floors.add(id: 'CeilingOther',
+    hpxml.frame_floors.add(id: 'CeilingBelowOther',
                            exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 900,
+                           interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                            insulation_assembly_r_value: 2.1,
                            other_space_above_or_below: HPXML::FrameFloorOtherSpaceAbove)
   elsif ['base-bldgtype-multifamily-adjacent-to-other-housing-unit.xml'].include? hpxml_file
@@ -4621,16 +4693,17 @@ def set_hpxml_frame_floors(hpxml_file, hpxml)
     hpxml.frame_floors[1].exterior_adjacent_to = HPXML::LocationOtherHeatedSpace
     hpxml.frame_floors[1].other_space_above_or_below = HPXML::FrameFloorOtherSpaceAbove
   elsif ['invalid_files/invalid-facility-type-surfaces.xml'].include? hpxml_file
-    hpxml.frame_floors.add(id: 'FloorOther',
+    hpxml.frame_floors.add(id: 'FloorAboveOther',
                            exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 900,
                            insulation_assembly_r_value: 2.1,
                            other_space_above_or_below: HPXML::FrameFloorOtherSpaceBelow)
-    hpxml.frame_floors.add(id: 'CeilingOther',
+    hpxml.frame_floors.add(id: 'CeilingBelowOther',
                            exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
                            interior_adjacent_to: HPXML::LocationLivingSpace,
                            area: 900,
+                           interior_finish_type: HPXML::InteriorFinishGypsumBoard,
                            insulation_assembly_r_value: 2.1,
                            other_space_above_or_below: HPXML::FrameFloorOtherSpaceAbove)
   end